requirements.txt
beautifulsoup4==4.12.3
pyotp==2.9.0
python-dotenv==1.0.1
selenium==4.21.0
